Login Page Complete
Login page is complete, it will check the DB and verify that the email and password is correct. Also added print functions in the dbpdo.php for all three DBs(printStaff, printProfessors, print(Books).

Changed "username" -> "email'

Added staff_index.php and prof_index.php to solidify the "fork" in the road for each end. Those assigned to either Professors or Staff should proceed with the respective index.